Bugfixes
*Fixed a gear connection issue causing massive lagspikes with certain bizarre setups
*Fixed the enchant effect turning stacks of items into 1 item
*Fixed light effect affecting blocks it shouldn't effect
*Fixed end-point blocks being linking start points, despite being invalid in all cases
*Fixed flux nodes ignoring player facing on placement
*Fixed beam flux stabilizers not accepting beams
*New master axis textures (courtesy of HarbingerOfCrazy)

+++ src/main/java/com/Da_Technomancer/crossroads/API/effects/LightEffect.java
@@ -1,5 +1,6 @@
 package com.Da_Technomancer.crossroads.API.effects;
 
+import com.Da_Technomancer.crossroads.API.beams.IBeamTransparent;
 import com.Da_Technomancer.crossroads.blocks.ModBlocks;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
@@ -14,16 +15,16 @@ public class LightEffect implements IEffect{
 	public void doEffect(World worldIn, BlockPos pos, int mult, EnumFacing dir){
 		IBlockState state = worldIn.getBlockState(pos);
 		BlockPos offsetPos;
-		if(state.getBlock() == ModBlocks.blockPureQuartz){
+		if(state.getBlock() == ModBlocks.blockPureQuartz && !(state.getBlock() instanceof IBeamTransparent)){
 			worldIn.setBlockState(pos, ModBlocks.blockLuminescentQuartz.getDefaultState());
-		}else if(state.getMaterial() == Material.ROCK && state.getBlock() != ModBlocks.blockLuminescentQuartz){
+		}else if(state.getMaterial() == Material.ROCK && state.getBlock() != ModBlocks.blockLuminescentQuartz && !(state.getBlock() instanceof IBeamTransparent)){
 			worldIn.setBlockState(pos, Blocks.GLOWSTONE.getDefaultState());
-		}else if(state.getMaterial() == Material.GLASS && state.getBlock() != Blocks.GLOWSTONE && state.getBlock() != ModBlocks.lightCluster){
+		}else if(state.getMaterial() == Material.GLASS && state.getBlock() != Blocks.GLOWSTONE && state.getBlock() != ModBlocks.lightCluster && !(state.getBlock() instanceof IBeamTransparent)){
 			worldIn.setBlockState(pos, Blocks.SEA_LANTERN.getDefaultState());
-		}else if(state.getBlock().isReplaceable(worldIn, pos)){
+		}else if(state.getBlock().isReplaceable(worldIn, pos) && !(state.getBlock() instanceof IBeamTransparent)){
 			worldIn.destroyBlock(pos, true);
 			worldIn.setBlockState(pos, ModBlocks.lightCluster.getDefaultState());
-		}else if(dir != null && worldIn.getBlockState(offsetPos = pos.offset(dir)).getBlock().isReplaceable(worldIn, offsetPos)){
+		}else if(dir != null && (state = worldIn.getBlockState(offsetPos = pos.offset(dir))).getBlock().isReplaceable(worldIn, offsetPos) && state.getBlock() != ModBlocks.lightCluster && !(state.getBlock() instanceof IBeamTransparent)){
 			worldIn.destroyBlock(offsetPos, true);
 			worldIn.setBlockState(offsetPos, ModBlocks.lightCluster.getDefaultState());
 		}
